% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cached OBR files}
\usage{
clear_cache()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Deletes all files downloaded and cached by the obr package. The next
function call will re-download fresh data from the OBR website.
}
\examples{
\donttest{
op <- options(obr.cache_dir = tempdir())
clear_cache()
options(op)
}

}
\concept{data access}
