% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintages.R
\name{obr_as_of}
\alias{obr_as_of}
\title{Find the OBR publication that was current on a given date}
\usage{
obr_as_of(date, publication = "EFO")
}
\arguments{
\item{date}{A Date, or anything coercible to one (e.g. \code{"2024-11-15"}).}

\item{publication}{Currently only \code{"EFO"} is supported.}
}
\value{
A length-one character vector containing the vintage label,
e.g. \code{"October 2024"}.
}
\description{
Returns the most recent EFO that had been published on or before \code{date}.
Useful for reproducing analyses as they would have looked at a given point
in time, before subsequent forecast revisions.
}
\examples{
obr_as_of("2024-11-15")
obr_as_of(Sys.Date())

}
\seealso{
Other vintage: 
\code{\link{obr_efo_vintages}()},
\code{\link{obr_pin}()},
\code{\link{obr_pinned}()},
\code{\link{obr_unpin}()}
}
\concept{vintage}
