% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vintages.R
\name{obr_pinned}
\alias{obr_pinned}
\title{Show the currently pinned OBR EFO vintage}
\usage{
obr_pinned()
}
\value{
Either a length-one character vector or \code{NULL}.
}
\description{
Returns the vintage string currently active via \code{\link[=obr_pin]{obr_pin()}}, or \code{NULL}
if no pin is set.
}
\examples{
obr_pinned()

}
\seealso{
Other vintage: 
\code{\link{obr_as_of}()},
\code{\link{obr_efo_vintages}()},
\code{\link{obr_pin}()},
\code{\link{obr_unpin}()}
}
\concept{vintage}
